clients/article: demo with reactive component in browsers
